Allow tests' set_up*() methods in snake case
This creates a new sniff, based on `Squiz.NamingConventions.ValidFunctionName`, that allows some methods to be ignored. This sniff checks that our methods are in camelCase but allows a given list of names in snake_case.
I was aiming for this list to be customizable within `ruleset.xml` but for an unknown reason, this doesn't seem to work, so the method names are directly listed in the sniff, in a public class property.
As a result, this sniff is used only in the "tests" folder, and the original sniff everywhere else.
The methods to ignore are: `set_up_before_class()`, `set_up()`, `assert_pre_conditions()`, `assert_post_conditions()`, `tear_down()`, and `tear_down_after_class()`.